Extract engines to separate object
