+++ app/com/codelab27/cards9/models/players/Match.scala
@@ -1,10 +1,17 @@
 package com.codelab27.cards9.models.players
 
-import com.codelab27.cards9.models.boards.{ Board, Red, Blue }
-import com.codelab27.cards9.models.boards.Board._
+import com.codelab27.cards9.models.boards.Board
 import com.codelab27.cards9.models.cards.Fight
 
-case class Match(
+/**
+  * A cards match.
+  *
+  * @param player1 player one identifier
+  * @param player2 player two identifier
+  * @param board the board
+  * @param fights list of already computed fights
+  */
+final case class Match(
   player1: Player.Id, // Red player
   player2: Player.Id, // Blue player
   board: Board,
@@ -12,6 +19,7 @@ case class Match(
 }
 
 object Match {
+
   case class Id(value: String) extends AnyVal
 
   case class RedScore(value: Int) extends AnyVal
@@ -20,27 +28,4 @@ object Match {
 
   case class Score(red: RedScore, blue: BlueScore)
 
-  /**
-    * Adds a new fight to the match.
-    *
-    * @param theMatch the match that receives the new fight
-    * @param fight the new fight
-    *
-    * @return match with the fight added
-    */
-  def addFight(theMatch: Match, fight: Fight): Match = theMatch.copy(fights = theMatch.fights :+ fight)
-
-  /**
-    * Get the current score of the match.
-    *
-    * @param theMatch match from where the score is extracted
-    *
-    * @return the current score
-    */
-  def score(theMatch: Match): Score = {
-    val redScore = RedScore(cardsOf(theMatch.board, Red).length)
-    val blueScore = BlueScore(cardsOf(theMatch.board, Blue).length)
-
-    Score(redScore, blueScore)
-  }
 }
