+++ app/com/codelab27/cards9/models/players/Player.scala
@@ -2,7 +2,13 @@ package com.codelab27.cards9.models.players
 
 import com.codelab27.cards9.models.cards.Card
 
-case class Player(
+/**
+  * A game player.
+  *
+  * @param id unique identifier
+  * @param cards cards owned by the player
+  */
+final case class Player(
   id: Player.Id,
   cards: List[Card.Id])
