+++ app/com/codelab27/cards9/utils/GameError.scala
@@ -1,0 +1,5 @@
+package com.codelab27.cards9.utils
+
+sealed trait GameError
+
+case class FightError(msg: String) extends GameError
