+++ build.sbt
@@ -12,6 +12,7 @@ scalacOptions ++= Seq(
   "-unchecked",
   "-deprecation",
   "-feature",
+  "-language:higherKinds",
   "-target:jvm-1.8",
   "-encoding", "UTF-8",
   "-Xfuture",
@@ -37,6 +38,7 @@ libraryDependencies ++= Seq(
     "com.typesafe.play" %% "play-slick" % "3.0.1",
     "com.typesafe.play" %% "play-slick-evolutions" % "3.0.1",
     "org.postgresql" % "postgresql" % "42.1.4",
+    "com.h2database" % "h2" % "1.4.196",
     "org.scalactic" %% "scalactic" % "3.0.4",
     "org.scalatest" %% "scalatest" % "3.0.4" % "test",
     "org.scalacheck" %% "scalacheck" % "1.13.4" % "test",
